+++ packages/core/src/index.ts
@@ -3,3 +3,4 @@ export * from "./filters";
 export * from "./search";
 export * from "./evolution-chain";
 export * from "./form-resolver";
+export * from "./type-chart";